# scheduling-worker/documentation/openapi.yaml
components:
  schemas:
    ActivityEvent:
      properties: {}
      type: object
    ActivityPattern:
      properties: {}
      type: object
...
    AnalyzeTimeUsageDto:
      properties: {}
      type: object
...
    TrackCompletionDto:
      properties: {}
      type: object
...
  securitySchemes:
    jwt:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Goal management, habit tracking, calendar integration, and ML-powered scheduling
  title: Scheduling Worker API
  version: 1.0.0
openapi: 3.0.0
paths:
  /activity/analyze:
    post:
      operationId: ActivityController_analyzeTimeUsage
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeTimeUsageDto'
        required: true
      responses:
        '200':
          description: Time usage analysis
      summary: Analyze time usage
      tags:
        - Activity
...
  /activity/efficiency/{userId}:
    get:
      operationId: ActivityController_getEfficiency
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: startDate
          required: true
          schema:
            type: string
        - in: query
          name: endDate
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schedule efficiency metrics
      summary: Get schedule efficiency metrics
      tags:
        - Activity
  /activity/events/{eventId}/complete:
    post:
      operationId: ActivityController_trackEventCompletion
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackCompletionDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityEvent'
          description: Event completion tracked
      summary: Track event completion
      tags:
        - Activity
  /activity/insights/{userId}:
    get:
      operationId: ActivityController_getInsights
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Productivity insights
      summary: Get productivity insights
      tags:
        - Activity
  /activity/patterns/{userId}:
    get:
      operationId: ActivityController_getPatterns
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityPattern'
          description: User activity patterns
        '404':
          description: Pattern not found
      summary: Get user activity patterns
      tags:
        - Activity
...
  /activity/predict/{eventId}:
    get:
      operationId: ActivityController_predictEventSuccess
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event success probability (0-1)
      summary: Predict event success probability
      tags:
        - Activity
...
  /scheduling/calendar/agenda:
    get:
      operationId: CalendarViewController_getAgendaView
      parameters:
        - in: query
          name: startDate
          required: true
          schema:
            type: string
        - in: query
          name: endDate
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agenda view data retrieved
      summary: Get agenda view data
      tags:
        - Calendar View
  /scheduling/calendar/day:
    get:
      operationId: CalendarViewController_getDayView
      parameters:
        - description: Date in YYYY-MM-DD format
          in: query
          name: date
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Day view data retrieved
      summary: Get day view data
      tags:
        - Calendar View
  /scheduling/calendar/month:
    get:
      operationId: CalendarViewController_getMonthView
      parameters:
        - in: query
          name: year
          required: true
          schema:
            type: number
        - in: query
          name: month
          required: true
          schema:
            type: number
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Month view data retrieved
      summary: Get month view data
      tags:
        - Calendar View
  /scheduling/calendar/week:
    get:
      operationId: CalendarViewController_getWeekView
      parameters:
        - description: Start date in YYYY-MM-DD format
          in: query
          name: startDate
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Week view data retrieved
      summary: Get week view data
      tags:
        - Calendar View
...
servers: []
...